+++ CHANGELOG.md
@@ -3,7 +3,16 @@
  - **BREAKING**: parameters of method `createDeployKey` have been modified.
 
 ## v5.3.0
- - Release.targetCommitsh is now deprecated. Use targetCommitish instead.
+- Add the ability to upload release assets.
+- Add the ability to get an existing release by tag name. 
+
+Deprecations: 
+- The `draft` and `prerelease` properties in the CreateRelease and Release 
+- classes have been renamed to `isDraft` and `isPrerelease` for clarity.
+- Release.targetCommitsh has been renamed to Release.targetCommitish. 
+- The `release` parameter in RepositoriesService.createRelease 
+has been renamed to `createRelease`.
+- `RepositoriesService.getRelease` has been renamed to `RepositoriesService.getReleaseById`
 
 ## v5.2.0
  - Add access to labels on Pull Requests https://github.com/DirectMyFile/github.dart/pull/163